WorkforceSystem/requirements.txt
mysql-connector-python>=9.0.0
matplotlib>=3.9.0
customtkinter>=5.2.2
